cleaned up Product Suppliers
